build: expose test while running dev server
