+++ package.json
@@ -40,7 +40,6 @@
     "qunitjs": "^2.4.1",
     "rollup": "^2.34.2",
     "rollup-plugin-copy": "^3.3.0",
-    "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-istanbul": "^2",
     "rollup-plugin-serve": "^1.1.0",
     "rollup-plugin-terser": "^7.0.2",
